Created new components and started fixing routes
