+++ front-end/src/App.js
@@ -1,5 +1,5 @@
-import React, {Component} from 'react';
-import {Link, Route} from 'react-router-dom';
+import React, { Component } from 'react';
+import { Link, Route } from 'react-router-dom';
 // import logo from './logo.svg'; import './App.css';
 import Home from './components/Home';
 import Clicker from './components/Clicker';
@@ -8,13 +8,12 @@ class App extends Component {
   render() {
     return (
       <div>
-        <nav>
+        {/* <nav>
           <Link to='/'>Login</Link>{' '}
-        </nav>
+        </nav> */}
         <div>
-          <Route exact path='/' component={Home}/>
-          <Route exact path='/clicker' component={Clicker}/>
-
+          <Route exact path='/' component={Home} />
+          <Route path='/clicker' component={Clicker} />
         </div>
       </div>
     );
